Split source files to improve compilation times.
